Move Feed UI Composition details from `FeedFetcheriOS` to `FeedFetchApp`
Move all composition details to the main module where the composition root lives
