+++ FeedFetchApp/FeedFetchApp/FeedImageDataLoaderPresentationAdapter.swift
@@ -3,6 +3,7 @@
 
 import Foundation
 import FeedFetcher
+import FeedFetcheriOS
 
 final class FeedImageDataLoaderPresentationAdapter<View: FeedImageView, Image>: FeedImageCellControllerDelegate where View.Image == Image {
     private let model: FeedImage
